refactor(steps): css import paths
