+++ packages/steps/package.json
@@ -7,6 +7,10 @@
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
   "files": [
+    "react.d.ts",
+    "react.mjs",
+    "react.js",
+    "styles.css",
     "src/*",
     "dist/*",
     "!src/*/__stories__/*",
@@ -26,7 +30,8 @@
     "test:watch": "yarn test -- --watchAll"
   },
   "sideEffects": [
-    "**/*.css"
+    "**/*.css",
+    "src/index.ts"
   ],
   "style": "dist/esm/css/index.css",
   "types": "dist/esm/index.d.ts",
