+++ packages/steps/src/index.ts
@@ -1,2 +1,4 @@
-export { default } from './react/index'
+import './css/index.css'
+import Steps from './react/index'
+export default Steps
 export * from './vars/index'
