+++ packages/steps/src/react/__stories__/component-Marker.story.tsx
@@ -3,7 +3,7 @@ import { Meta, Story } from '@storybook/react/types-6-0'
 import React from 'react'
 
 import Context, { initialContext } from '../context'
-import Steps from '../index'
+import Steps from '../../index'
 import * as vars from '../../vars/index'
 
 const { Marker } = Steps
