+++ packages/steps/src/react/__stories__/component-Step.story.tsx
@@ -4,7 +4,7 @@ import { Meta, Story } from '@storybook/react/types-6-0'
 import React from 'react'
 
 import Context, { initialContext } from '../context'
-import Steps from '../index'
+import Steps from '../../index'
 import { FixedWidthContainer } from './shared'
 
 const { Step } = Steps
