+++ packages/steps/src/react/__stories__/examples.story.tsx
@@ -3,7 +3,7 @@ import { Meta, Story } from '@storybook/react/types-6-0'
 import React from 'react'
 
 import * as vars from '../../vars/index'
-import Steps from '../index'
+import Steps from '../../index'
 
 import { FixedWidthContainer } from './shared'
