+++ packages/steps/src/react/__stories__/vertical.story.tsx
@@ -1,7 +1,7 @@
 import { Meta, Story } from '@storybook/react/types-6-0'
 import React from 'react'
 
-import Steps from '../index'
+import Steps from '../../index'
 import { FixedWidthContainer, useStoryData } from './shared'
 
 export default {
