+++ packages/steps/src/react/index.tsx
@@ -6,7 +6,6 @@ import {
 import { useTheme } from '@pluralsight/ps-design-system-theme'
 import React from 'react'
 
-import '../css/index.css'
 import * as vars from '../vars/index'
 
 import Context, { ContextValue, initialContext } from './context'
